refactor: 트랜잭션 메서드와 일반 메서드의 클래스 분리
