+++ src/main/java/com/example/WonkaoTalk/common/oauth/OAuthRevocationClient.java
@@ -2,16 +2,11 @@ package com.example.WonkaoTalk.common.oauth;
 
 import com.example.WonkaoTalk.common.exception.BusinessException;
 import com.example.WonkaoTalk.common.exception.ErrorCode;
-import com.example.WonkaoTalk.domain.auth.entity.AuthSocial;
-import com.example.WonkaoTalk.domain.auth.entity.OAuthRevocationFailure;
 import com.example.WonkaoTalk.domain.auth.repo.AuthSocialRepo;
-import com.example.WonkaoTalk.domain.auth.repo.OAuthRevocationFailureRepo;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Propagation;
-import org.springframework.transaction.annotation.Transactional;
 
 @Slf4j
 @Component
@@ -20,7 +15,7 @@ public class OAuthRevocationClient {
 
   private final List<OAuthRevocationProvider> revocationProviders;
   private final AuthSocialRepo authSocialRepo;
-  private final OAuthRevocationFailureRepo failureRepo;
+  private final OAuthRevocationFailureProcessor failureProcessor;
 
   public void revokeIfSocialAccountExists(Long authId) {
     authSocialRepo.findByAuthId(authId).ifPresent(social -> {
@@ -35,18 +30,8 @@ public class OAuthRevocationClient {
       } catch (Exception e) {
         log.error("소셜 연동 해제 통신 실패. Outbox에 실패 이력을 기록합니다. AuthId: {}", authId, e);
         // Fallback: 실패 이력 저장 (이후 스케줄러가 재시도)
-        saveFailureEvent(social);
+        failureProcessor.saveFailureEvent(social);
       }
     });
-  }
-
-  @Transactional(propagation = Propagation.REQUIRES_NEW)
-  protected void saveFailureEvent(AuthSocial social) {
-    OAuthRevocationFailure failure = OAuthRevocationFailure.builder()
-        .provider(social.getProvider())
-        .providerUserId(social.getProviderUserId())
-        .providerRefreshToken(social.getProviderRefreshToken())
-        .build();
-    failureRepo.save(failure);
   }
 }
